requirements.txt
requests
beautifulsoup4
cryptography
python-dotenv
spacy